New: “post” pages
Post pages are title-less bits of content, kind of like tweets or Facebook statuses. Because they lack a title and can be related to any project, group, or anything else on the site, they’re located at website.com/##########.

This tweet-like format will eventually be the primary way users will be able to post to the site via a webapp. Related content can be tagged as usual, allowing project pages to contain a running dialogue of comments and ideas.

Maybe related pages can be tagged in a hashtag-like fashion?