Fixed post images not being saved correctly
When I made the change to do thumbnail processing on the backend and return the thumb URL to ContentTools, I neglected to account for that URL being a bit different with the image dimensions like ‘-700x340’ appended to the end.

A tweak of editor.js fixes that, allowing images to be saved correctly once again.

I also made the menu only draggable when viewing the settings page.