Major folder restructuring
In order to make Acorn updates easier in the future, a number of changes had to be made.

Assets are now stored within the “site” folder, except for the “avatars” subfolder, which is now within a new top-level “users” folder, along with the “accounts” folder that used to be within “site”.

“Sessions”, “thumbs” and the HTML “cache” folder have been moved from outside of “site” into a new top-level “cache” folder.

This new structure feels relatively simpler and cleaner. Cache has temporary stuff, content is the bulk of the site and shouldn’t be touched, kirby is the system folder, site has everything Acorn-related, and users contains all user-specific things that shouldn’t be touched either.