Site-wide background color support
Setting the site’s bg-color-primary within site.txt now changes the background color of the entire site.

The color needs to be inputted as an RGB triplet like “240, 240, 240” to allow alpha channels to be appended to the CSS variable that stores the RGB value.

This is a workaround for the lack of any opacity setting for the background-color CSS property, discussed here: https://stackoverflow.com/a/41265350