Changed ‘users’ folder location
Because example.com/users would be a common location for a user directory, the users folder containing account info and avatars needed to be moved elsewhere. For now it’s in a new top-level ‘acorn’ folder. Ideally it would be within the site folder, but that would make “upgrading” site site manually more difficult unless the entirety of Acorn were a series of Kirby plugins.