New content update mechanism
The new acorn-update.php file handles updating/upgrading page.txt files to conform to the latest structure required of the current Acorn version. Old methods/functions are moved to the methods-deprecated.php file to be used one last time by the update process. Every page’s modification date is checked against the time that the newest version of Acorn was installed to determine if an update is necessary. The entire site can be updated at once by using the /flush route to “touch” and check every page.