Moved content snippet logic to content tag
Calling a snippet forced content cards to appear before the page’s text, which wasn’t the intent. By moving all logic into the tag, content sections can be displayed anywhere on the page and formatted accordingly. I may have to reconsider this in the future if I want content to be callable via an API.