Fixed “Default” color setting visual feedback
Previously, a page with a color setting of “default” would display whatever that color was as the default option within the Page Color settings field. This made it very easy for the user to accidentally save the page as “Green” when they really just wanted it to obey whatever their site-wide default was. Now the “Default” option is treated just like any other color and displayed appropriately.